Extracted: RideHailRequestType, RideHailRequest and RideHailResponse
