+++ src/main/scala/beam/agentsim/agents/ridehail/RideHailManager.scala
@@ -1054,35 +1054,6 @@ object RideHailManager {
 
   case class NotifyIterationEnds()
 
-  sealed trait RideHailRequestType
-
-  case object RideHailInquiry extends RideHailRequestType
-
-  case object ReserveRide extends RideHailRequestType
-
-  case class RideHailRequest(
-    requestType: RideHailRequestType,
-    customer: VehiclePersonId,
-    pickUpLocation: Location,
-    departAt: BeamTime,
-    destination: Location
-  ) {
-    // We make requestId be independent of request type, all that matters is details of the customer
-    lazy val requestId: Int =
-      this.copy(requestType = RideHailInquiry).hashCode()
-  }
-
-  object RideHailRequest {
-
-    val dummy = RideHailRequest(
-      RideHailInquiry,
-      VehiclePersonId(Id.create("dummy", classOf[Vehicle]), Id.create("dummy", classOf[Person])),
-      new Coord(Double.NaN, Double.NaN),
-      DiscreteTime(Int.MaxValue),
-      new Coord(Double.NaN, Double.NaN)
-    )
-  }
-
   case class TravelProposal(
     rideHailAgentLocation: RideHailAgentLocation,
     timeToCustomer: Long,
@@ -1091,20 +1062,6 @@ object RideHailManager {
     responseRideHail2Pickup: RoutingResponse,
     responseRideHail2Dest: RoutingResponse
   )
-
-  case class RideHailResponse(
-    request: RideHailRequest,
-    travelProposal: Option[TravelProposal],
-    error: Option[ReservationError] = None,
-    triggersToSchedule: Vector[ScheduleTrigger] = Vector()
-  )
-
-  object RideHailResponse {
-    val dummy = RideHailResponse(RideHailRequest.dummy, None, None)
-
-    def dummyWithError(error: ReservationError) =
-      RideHailResponse(RideHailRequest.dummy, None, Some(error))
-  }
 
   private case class RoutingResponses(
     request: RideHailRequest,
