+++ src/main/scala/beam/agentsim/agents/ridehail/RideHailRequestType.scala
@@ -1,0 +1,7 @@
+package beam.agentsim.agents.ridehail
+
+sealed trait RideHailRequestType
+
+case object RideHailInquiry extends RideHailRequestType
+
+case object ReserveRide extends RideHailRequestType
